requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.40.1
requests==2.31.0
python-dotenv==1.0.1

langchain-core==0.3.72
langchain-huggingface==0.3.1
huggingface-hub==0.33.4
tokenizers>=0.19.1
